requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
alembic
python-dotenv
mysql-connector-python